# swagger.yml
openapi: 3.0.2
info:
  title: booktrack
  version: 1.0.0
paths:
  /users:
    summary: Path used to manage the list of users.
    description: >-
      The REST endpoint/path used to list and create zero or more `User` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `User` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      tags:
        - /users
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                'Invalid email: {reason}':
                  value: 'Invalid email: {reason}'
                Email and password are required:
                  value: Email and password are required
                Password must be at least 8 characters long:
                  value: Password must be at least 8 characters long
                Email must be in a valid format:
                  value: Email must be in a valid format
          description: Client side error
        '409':
          content:
            text/html:
              examples:
                Email already exists:
                  value: Email already exists
          description: The email address provided already exists in the system.
      operationId: createUser
      summary: Create a User
      description: Creates a new instance of a `User`.
  /sessions:
    summary: Path used to manage the list of sessions.
    description: >-
      The REST endpoint/path used to list and create zero or more `Session` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `Session` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
        required: true
      tags:
        - /sessions
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Email and password are required:
                  value: Email and password are required
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Invalid password:
                  value: Invalid password
          description: Unauthorized
        '404':
          content:
            text/html:
              examples:
                User not found:
                  value: User not found
          description: Not Found
      operationId: createSession
      summary: Create a Session
      description: 'Creates a new instance of a `Session`, signing the user in. Returns session ID.'
    delete:
      tags:
        - /sessions
      responses:
        '204':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      security:
        - Bearer: [ ]
      summary: Destroy a Session
      description: 'Destroys a `Session`, signing the user out. Reads the session ID from Authorization header.'
  /books:
    summary: Path used to manage the list of books.
    description: >-
      The REST endpoint/path used to list and create zero or more `Book` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    get:
      tags:
        - /books
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
          description: Successful response - returns an array of `Book` entities.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      security:
        - Bearer: [ ]
      operationId: getBooktrack
      summary: List All Books
      description: Gets a list of all `Book` entities.
    post:
      requestBody:
        description: A new `Book` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
            examples:
              Book:
                value:
                  content: text
                  title: text
        required: true
      tags:
        - /books
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      security:
        - Bearer: [ ]
      operationId: addBook
      summary: Add a Book
      description: Creates a new instance of a `Book`.
  '/books/:id':
    summary: Path used to manage the list of books.
    description: >-
      The REST endpoint/path used to edit `Book` entities.  This path contains a `PUT` operation to
      perform the edit books.
    put:
      requestBody:
        description: A `Book` to be edited.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
            examples:
              Book:
                value:
                  content: text
                  title: text
        required: true
      tags:
        - '/books/:id'
      responses:
        '400':
          description: Bad request (Title and content are required)
        '401':
          description: Unauthorized
        '404':
          description: Book ID not found.
      deprecated: false
      security:
        - Bearer: [ ]
      operationId: editBook
      summary: Edit Book
      description: Updates an instance of `Book`.

    delete:
      tags:
        - '/books/:id'
      responses:
        '204':
          description: Successful response - delete book
        '401':
          content:
            text/xml:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
        '404':
          description: Book not found
      deprecated: false
      security:
        - Bearer: [ ]
      operationId: deleteBook
      summary: Delete the book
      description: Deletes an instance of `Book`.
components:
  schemas:
    User:
      title: Root Type for User
      description: ''
      type: object
      properties:
        id:
          format: int32
          description: A field that represents the unique identifier of the user.
          type: integer
          example: '12345'
        email:
          description: A field that represents the email address of the user.
          type: string
          example: john.doe@example.com
        password:
          description: A field that represents the password of the user.
          type: string
          example: MySecurePassword123
      example:
        id: 3
        email: email@example.com
        password: pass123
    Session:
      title: Root Type for Session
      description: ''
      type: object
      properties:
        id:
          type: string
        userId:
          format: int32
          type: integer
      example:
        id: 330d67de-432b-48ce-a617-0bd6774ebcee
        userId: 1
    Credentials:
      title: Root Type for Credentials
      description: ''
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: admin
        password: KollneKollne
    Book:
      title: Root Type for Book
      description: ''
      type: object
      properties:
        content:
          type: string
        id:
          format: int32
          type: integer
        title:
          type: string
        userId:
          format: int32
          type: integer
      example:
        content: This is the content of book 1
        id: 1
        title: Book 1
        userId: 1
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
tags:
  - name: /users
    description: ''
  - name: /sessions
    description: ''
  - name: /books
    description: ''
  - name: '/books/:id'
    description: ''